🧹 Deduplicate parseReadyCount into shared lib/k8s.ts
Fixes #10135

Five card helper files (crio, kubevirt, ovn, kubeflex, k3s) had
byte-identical copies of parseReadyCount. Extract to lib/k8s.ts
and re-export from each helper so existing test imports still work.
