+++ web/src/components/cards/multi-tenancy/kubevirt-status/helpers.ts
@@ -66,18 +66,8 @@ export function isVirtLauncher(labels?: Record<string, string>): boolean {
   return appLabel === VIRT_LAUNCHER_LABEL
 }
 
-/**
- * Parse a Kubernetes "ready" string like "1/1" into numeric values.
- */
-export function parseReadyCount(ready?: string): { ready: number; total: number } {
-  const [readyPart, totalPart] = String(ready ?? '').split('/')
-  const readyCount = Number.parseInt(readyPart, 10)
-  const totalCount = Number.parseInt(totalPart, 10)
-  return {
-    ready: Number.isFinite(readyCount) ? readyCount : 0,
-    total: Number.isFinite(totalCount) ? totalCount : 0,
-  }
-}
+import { parseReadyCount } from '../../../../lib/k8s'
+export { parseReadyCount } from '../../../../lib/k8s'
 
 /**
  * Determine whether a pod is healthy based on its status and ready ratio.
