+++ web/src/components/cards/multi-tenancy/ovn-status/helpers.ts
@@ -5,6 +5,9 @@
  * pod and annotation data into the metrics shown by the card.
  */
 
+import { parseReadyCount } from '../../../../lib/k8s'
+export { parseReadyCount } from '../../../../lib/k8s'
+
 // ============================================================================
 // Types
 // ============================================================================
@@ -86,19 +89,6 @@ export function isPodHealthy(pod: OvnPodInfo): boolean {
 
   const { ready, total } = parseReadyCount(pod.ready)
   return total > 0 && ready === total
-}
-
-/**
- * Parse a Kubernetes "ready" string like "1/1" into numeric values.
- */
-export function parseReadyCount(ready?: string): { ready: number; total: number } {
-  const [readyPart, totalPart] = String(ready ?? '').split('/')
-  const readyCount = Number.parseInt(readyPart, 10)
-  const totalCount = Number.parseInt(totalPart, 10)
-  return {
-    ready: Number.isFinite(readyCount) ? readyCount : 0,
-    total: Number.isFinite(totalCount) ? totalCount : 0,
-  }
 }
 
 /**
